+++ extension.js
@@ -23,6 +23,7 @@ import {WindowPicker} from './src/WindowPicker.js';
 import * as utils from './src/utils.js';
 
 import Apparition from './src/effects/Apparition.js';
+import AuraGlow from './src/effects/AuraGlow.js';
 import BrokenGlass from './src/effects/BrokenGlass.js';
 import Doom from './src/effects/Doom.js';
 import EnergizeA from './src/effects/EnergizeA.js';
@@ -34,17 +35,19 @@ import Glitch from './src/effects/Glitch.js';
 import Hexagon from './src/effects/Hexagon.js';
 import Incinerate from './src/effects/Incinerate.js';
 import Matrix from './src/effects/Matrix.js';
+import Mushroom from './src/effects/Mushroom.js';
 import PaintBrush from './src/effects/PaintBrush.js';
 import Pixelate from './src/effects/Pixelate.js';
 import PixelWheel from './src/effects/PixelWheel.js';
 import PixelWipe from './src/effects/PixelWipe.js';
 import Portal from './src/effects/Portal.js';
+import RGBWarp from './src/effects/RGBWarp.js';
 import SnapOfDisintegration from './src/effects/SnapOfDisintegration.js';
+import TeamRocket from './src/effects/TeamRocket.js';
 import TRexAttack from './src/effects/TRexAttack.js';
 import TVEffect from './src/effects/TVEffect.js';
 import TVGlitch from './src/effects/TVGlitch.js';
 import Wisps from './src/effects/Wisps.js';
-
 
 import * as Main from 'resource:///org/gnome/shell/ui/main.js';
 import {Workspace} from 'resource:///org/gnome/shell/ui/workspace.js';
@@ -71,14 +74,13 @@ export default class BurnMyWindows extends Extension {
 
     // New effects must be registered here and in prefs.js.
     this._ALL_EFFECTS = [
-      new Apparition(), new BrokenGlass(), new Doom(),
-      new EnergizeA(),  new EnergizeB(),   new Fire(),
-      new Focus(),      new Glide(),       new Glitch(),
-      new Hexagon(),    new Incinerate(),  new Matrix(),
-      new PaintBrush(), new Pixelate(),    new PixelWheel(),
-      new PixelWipe(),  new Portal(),      new SnapOfDisintegration(),
-      new TRexAttack(), new TVEffect(),    new TVGlitch(),
-      new Wisps(),
+      new Apparition(), new AuraGlow(),   new BrokenGlass(), new Doom(),
+      new EnergizeA(),  new EnergizeB(),  new Fire(),        new Focus(),
+      new Glide(),      new Glitch(),     new Hexagon(),     new Incinerate(),
+      new Matrix(),     new PaintBrush(), new Pixelate(),    new PixelWheel(),
+      new PixelWipe(),  new Portal(),     new RGBWarp(),     new SnapOfDisintegration(),
+      new TeamRocket(), new TRexAttack(), new TVEffect(),    new TVGlitch(),
+      new Wisps(),      new Mushroom()
     ];
 
     // Load all of our resources.