+++ prefs.js
@@ -24,6 +24,7 @@ import * as utils from './src/utils.js';
 import {ProfileManager} from './src/ProfileManager.js';
 
 import Apparition from './src/effects/Apparition.js';
+import AuraGlow from './src/effects/AuraGlow.js';
 import BrokenGlass from './src/effects/BrokenGlass.js';
 import Doom from './src/effects/Doom.js';
 import EnergizeA from './src/effects/EnergizeA.js';
@@ -35,12 +36,15 @@ import Glitch from './src/effects/Glitch.js';
 import Hexagon from './src/effects/Hexagon.js';
 import Incinerate from './src/effects/Incinerate.js';
 import Matrix from './src/effects/Matrix.js';
+import Mushroom from './src/effects/Mushroom.js';
 import PaintBrush from './src/effects/PaintBrush.js';
 import Pixelate from './src/effects/Pixelate.js';
 import PixelWheel from './src/effects/PixelWheel.js';
 import PixelWipe from './src/effects/PixelWipe.js';
 import Portal from './src/effects/Portal.js';
+import RGBWarp from './src/effects/RGBWarp.js';
 import SnapOfDisintegration from './src/effects/SnapOfDisintegration.js';
+import TeamRocket from './src/effects/TeamRocket.js';
 import TRexAttack from './src/effects/TRexAttack.js';
 import TVEffect from './src/effects/TVEffect.js';
 import TVGlitch from './src/effects/TVGlitch.js';
@@ -71,10 +75,15 @@ export default class BurnMyWindowsPreferences extends ExtensionPreferences {
 
     // New effects must be registered here and in extension.js.
     this._ALL_EFFECTS = [
-      Apparition, BrokenGlass, Doom,       EnergizeA, EnergizeB,  Fire,
-      Focus,      Glide,       Glitch,     Hexagon,   Incinerate, Matrix,
-      PaintBrush, Pixelate,    PixelWheel, PixelWipe, Portal,     SnapOfDisintegration,
-      TRexAttack, TVEffect,    TVGlitch,   Wisps,
+      Apparition, AuraGlow,   BrokenGlass,
+      Doom,       EnergizeA,  EnergizeB,
+      Fire,       Focus,      Glide,
+      Glitch,     Hexagon,    Incinerate,
+      Matrix,     Mushroom,   PaintBrush,
+      Pixelate,   PixelWheel, PixelWipe,
+      Portal,     RGBWarp,    SnapOfDisintegration,
+      TeamRocket, TRexAttack, TVEffect,
+      TVGlitch,   Wisps,
     ];
 
 
@@ -183,22 +192,33 @@ export default class BurnMyWindowsPreferences extends ExtensionPreferences {
     this._searchEntry.connect('search-changed', () => {
       const query = this._searchEntry.get_text().toLowerCase();
 
-      this._effectRows.forEach(er => {
+      // If only one effect is left, we will expand it automatically.
+      let lastRow = null;
+      let numRows = 0;
+
+      this._effectRows.forEach(row => {
+        row.set_expanded(false);
+
         if (query === '') {
-          er.show();  // Show all effects if query is empty
+          row.show();  // Show all effects if query is empty
         } else {
           // Show or hide each effect based on query match
-          const showEffect = er.name.toLowerCase().includes(query);
-          showEffect ? er.show() : er.hide();
+          const showEffect = row.name.toLowerCase().includes(query);
 
-          // TODO
-          // maybe add a fuzzy search later
-          /*
-          const showEffect = utils.fuzzyMatch(er.name.toLowerCase(),
-          query.toLowerCase()); showEffect ? er.show() : er.hide();
-          */
+          if (showEffect) {
+            row.show();
+            lastRow = row;
+            numRows++;
+          } else {
+            row.hide();
+          }
         }
       });
+
+      // If only one effect is left, we expand it automatically.
+      if (numRows === 1) {
+        lastRow.set_expanded(true);
+      }
     });
 
     // Then add a preferences group for the effect expander rows.